feat: hide link to stream display table if no existing streams
outgoing streams are now preloaded on `App` to improve load time.
